requirements.txt
Flask==3.0.0
azure-storage-blob
pyodbc

gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
